fix #26: raise #UD on unsupported VMCALL
VMCALL raises #UD when a VMM is not installed. If a VMM is
installed, how to handle error cases is up to the VMM. All
major VMMs (KVM, VMware, VirtualBox at least) raise #UD on
unsupported VMCALL letting a guest behave as if there were
no VMM in the system.

This patch changes behaviour of the VMCALL handler to
inject #UD on unsupported VMCALL.

See handle_vmcall() for KVM, and hmR0VmxExitVmcall() for
VirtualBox implementation as references.