fix #31 (wrong memory type)
This update fixes a system hang and power callback issues seen on
Kaby Lake systems.

The root cause was that a memory type set for EPT PTE did not respect
MTRR and overwriting UC set in MTRR with WB, which led to WB memory
type for MMIO regions.

This update is to read and store all MTRRs at initialization, and
consult them when EPT entries are built.

Signed-off-by: Satoshi Tanda <tanda.sat@gmail.com>